refactor storage/range
- executeCmd and all commands (plus stuff mostly used from them):
  range_command.go
- raft storage implementation:
  range_raftstorage.go
- everything else: range.go
